refactor(Core/Creature): Split creature.h
